+++ examples/local_noise_simulation.py
@@ -11,8 +11,20 @@
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
 
-"""Version information.
-   Version number (major.minor.patch[-label])
-"""
+from braket.circuits import Circuit, Noise
+from braket.devices import LocalSimulator
 
-__version__ = "1.6.3.dev0"
+device = LocalSimulator("braket_dm")
+
+circuit = Circuit().x(0).x(1).bit_flip(0, probability=0.1)
+print("First example: ")
+print(circuit)
+print(device.run(circuit, shots=1000).result().measurement_counts)
+
+
+circuit = Circuit().x(0).x(1)
+noise = Noise.BitFlip(probability=0.1)
+circuit.apply_gate_noise(noise)
+print("Second example: ")
+print(circuit)
+print(device.run(circuit, shots=1000).result().measurement_counts)