+++ src/braket/aws/aws_quantum_task.py
@@ -460,7 +460,9 @@ def _(
     create_task_kwargs.update(
         {
             "action": problem.to_ir().json(),
-            "deviceParameters": device_params.json(),
+            "deviceParameters": DwaveDeviceParameters.parse_obj(device_parameters).json(
+                exclude_none=True
+            ),
         }
     )